api/requirements.txt
flask
PyMySQL[rsa]
flask-cors
python-dotenv
uwsgi
